requirements.txt
django~=4.2.13
pillow~=10.3.0
django-jazzmin~=3.0.0
pypdf2~=3.0.1 
reportlab~=4.2.2
django-ckeditor